+++ plugins/com.google.cloud.tools.eclipse.integration.appengine/META-INF/MANIFEST.MF
@@ -22,7 +22,7 @@ Require-Bundle: com.google.cloud.tools.eclipse.test.dependencies,
  com.google.cloud.tools.eclipse.appengine.standard.java8;bundle-version="0.1.0",
  com.google.cloud.tools.eclipse.appengine.localserver;bundle-version="0.1.0",
  com.google.cloud.tools.eclipse.appengine.facets;bundle-version="0.1.0"
-Import-Package: com.google.cloud.tools.appengine.operations.cloudsdk;version="0.7.3",
+Import-Package: com.google.cloud.tools.appengine.operations;version="0.7.3",
  com.google.cloud.tools.eclipse.appengine.compat.cte13,
  com.google.cloud.tools.eclipse.appengine.deploy,
  com.google.cloud.tools.eclipse.appengine.deploy.flex,
