+++ plugins/com.google.cloud.tools.eclipse.sdk.test/src/com/google/cloud/tools/eclipse/sdk/internal/CloudSdkPreferenceResolverTest.java
@@ -26,8 +26,8 @@ import static org.mockito.Mockito.when;
 
 import com.google.cloud.tools.appengine.AppEngineException;
 import com.google.cloud.tools.appengine.operations.CloudSdk;
-import com.google.cloud.tools.appengine.operations.CloudSdkNotFoundException;
-import com.google.cloud.tools.appengine.operations.CloudSdkResolver;
+import com.google.cloud.tools.appengine.operations.cloudsdk.CloudSdkNotFoundException;
+import com.google.cloud.tools.appengine.operations.cloudsdk.CloudSdkResolver;
 import java.io.File;
 import java.util.Collections;
 import java.util.List;
